+++ docs/release-notes.md
@@ -2,6 +2,7 @@
 * Fixed SiN launching with Zero Tier enabled on AMD GPUs.
 * Host name look-up implemented for games using WinSock2, which enables more games to work with the Zero Tier wrapper.
 * Fixed crash/issue launching partially supported games.
+* Zero Tier networking library now logs data which can be viewed by tools such as DebugView.
 
 # Siege Studio + Siege Launcher 26.23 Development Release
 * Various fixes and improvements for Medal of Honor - Allied Assault and it's expansions.